requirements.txt
av==12.0.0
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
coloredlogs==15.0.1
ctranslate2==4.2.1
faster-whisper==1.0.2
ffmpeg-python==0.2.0
filelock==3.14.0
flatbuffers==24.3.25
fsspec==2024.3.1
future==1.0.0
huggingface-hub==0.23.0
humanfriendly==10.0
idna==3.7
mpmath==1.3.0
numpy==1.26.4
onnxruntime==1.17.3
openpyxl==3.1.2
packaging==24.0
protobuf==5.26.1
pyreadline3==3.4.1
PyYAML==6.0.1
requests==2.31.0
sympy==1.12
tokenizers==0.19.1
tqdm==4.66.4
typing_extensions==4.11.0
urllib3==2.2.1